requirements.txt
alembic==1.12.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
Faker==18.3.1
fast-ml==3.68
fastjsonschema==2.16.3
filelock==3.12.0
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
fonttools==4.42.1
fqdn==1.5.1
greenlet==2.0.2
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
ipykernel==6.22.0
ipynb-py-convert==0.4.6
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
minio==7.1.16
mistune==2.0.5
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
onnx==1.14.1
onnxconverter-common==1.14.0
openpyxl==3.1.2
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.2
psutil==5.9.4
psycopg2==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.8.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
skl2onnx==1.15.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.9
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.4
virtualenv==20.22.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.7
WTForms==3.0.1
